requirements.txt
pynput
pyperclip
Pillow
pywin32
requests
websockets
